Refactor Modbus related classes and methods into a Modbus namespace.
As part of this change, we have moved Modbus classes (like FunctionCode, ExceptionCode), methods and utility functions into a new Modbus namespace. This makes the code easier to manage and segregates Modbus related code from the rest of the application. This will also facilitate easier future enhancements and potential reuse of the Modbus code.