Add register existence check before insert
A new function `registerExists` has been added to check if a register with the specified address already exists in a vector of registers. This function will throw an `invalid_argument` exception if an attempt is made to insert a register with an existing address. The change has also been covered in unit tests to ensure its correctness.