Define MBAP header length constant and modify related usages
Added a constant for the Modbus Application Protocol (MBAP) header length in ModbusServer.h. This constant is then utilized in ModbusServer.cpp to define the number of bytes for Protocol Data Unit (PDU). Also, the unitIdentifier is added correctly to the MBAP and tests are updated. In CMakeLists.txt, a commented include directory line is removed.